# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "HTTP Basic with the subscription token as the password. The username is ignored — tokens are self-identifying. This is the form `git clone http://user:\u003ctoken\u003e@host/git/repo` produces, so git-over-HTTP works out of the box.",
        "title": "GiGot API",
        "contact": {
            "name": "Peter van de Pas"
        },
        "license": {
            "name": "MIT"
        },
        "version": "0.1.0"
    },
    "host": "localhost:3417",
    "basePath": "/api",
    "paths": {
...
        "/admin/repos/{name}/destinations/{id}/sync": {
            "post": {
                "security": [
                    {
                        "SessionAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Invokes `git push +refs/heads/*:refs/heads/* +refs/audit/*:refs/audit/*`\nagainst the destination's URL, using the vault credential\nreferenced by `credential_name`. Runs synchronously — the\nresponse is the updated destination with `last_sync_at`,\n`last_sync_status`, and (on failure) `last_sync_error`\npopulated. On success the vault credential's `last_used`\ntimestamp is also touched. Destinations with enabled=false\nstill accept this call (manual is explicit operator intent);\nthe flag gates only the automatic post-receive fan-out.\n\nAuthorization mirrors the rest of /destinations*: admin\nroute requires an admin session; subscriber route requires\nbearer token + repo scope + admin/maintainer role +\n`mirror` ability (see accounts.md §6.1, remote-sync.md\n§2.6 and §5).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin",
                    "repos"
                ],
                "summary": "Trigger a manual mirror-sync push on one destination",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Repo name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Destination id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.DestinationView"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Subscriber route: missing mirror ability, regular role, or repo out of scope",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Repo, destination, or credential (deleted) not found",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Credential referenced by destination no longer exists in the vault",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/repos/{name}/destinations/{id}/sync": {
            "post": {
                "security": [
                    {
                        "SessionAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Invokes `git push +refs/heads/*:refs/heads/* +refs/audit/*:refs/audit/*`\nagainst the destination's URL, using the vault credential\nreferenced by `credential_name`. Runs synchronously — the\nresponse is the updated destination with `last_sync_at`,\n`last_sync_status`, and (on failure) `last_sync_error`\npopulated. On success the vault credential's `last_used`\ntimestamp is also touched. Destinations with enabled=false\nstill accept this call (manual is explicit operator intent);\nthe flag gates only the automatic post-receive fan-out.\n\nAuthorization mirrors the rest of /destinations*: admin\nroute requires an admin session; subscriber route requires\nbearer token + repo scope + admin/maintainer role +\n`mirror` ability (see accounts.md §6.1, remote-sync.md\n§2.6 and §5).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin",
                    "repos"
                ],
                "summary": "Trigger a manual mirror-sync push on one destination",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Repo name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Destination id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.DestinationView"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Subscriber route: missing mirror ability, regular role, or repo out of scope",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Repo, destination, or credential (deleted) not found",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Credential referenced by destination no longer exists in the vault",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "server.DestinationView": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "credential_name": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "last_sync_at": {
                    "type": "string"
                },
                "last_sync_error": {
                    "type": "string"
                },
                "last_sync_status": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "server.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "repository not found"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        },
        "BearerAuth": {
            "description": "Enter your bearer token as: Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "SessionAuth": {
            "description": "Session cookie minted by POST /api/admin/login (or any successful OAuth callback). Required by every /api/admin/* endpoint and the /fragments/* template server. Browsers attach it automatically; programmatic callers must include the cookie header on every request.",
            "type": "apiKey",
            "name": "gigot_session",
            "in": "cookie"
        }
    }
}